update: ディレクトリパスを@表記に変更
